update on SerializedDataset and save in a separate file
