+++ hydragnn/utils/pickledataset.py
@@ -144,70 +144,3 @@ class SimplePickleWriter:
                 dirfname = os.path.join(basedir, subdir, fname)
             with open(dirfname, "wb") as f:
                 pickle.dump(data, f)
-
-
-class SerializedDataset(AbstractBaseDataset):
-    """Serialized Dataset"""
-
-    def __init__(self, basedir, datasetname, label):
-        """
-        Parameters
-        ----------
-        basedir: basedir
-        datasetname: dataset name
-        label: label
-        """
-        super().__init__()
-
-        self.basedir = basedir
-        self.datasetname = datasetname
-        self.label = label
-
-        fname = os.path.join(basedir, "%s-%s.pkl" % (datasetname, label))
-        with open(fname, "rb") as f:
-            self.minmax_node_feature = pickle.load(f)
-            self.minmax_graph_feature = pickle.load(f)
-            self.dataset = pickle.load(f)
-
-        log("Pickle files:", self.label, len(self.dataset))
-
-    def len(self):
-        return len(self.dataset)
-
-    def get(self, i):
-        return self.dataset[i]
-
-
-class SerializedWriter:
-    """Serialized Dataset Writer"""
-
-    def __init__(
-        self,
-        dataset,
-        basedir,
-        datasetname,
-        label="total",
-        minmax_node_feature=None,
-        minmax_graph_feature=None,
-        comm=MPI.COMM_WORLD,
-    ):
-        """
-        Parameters
-        ----------
-        dataset: locally owned dataset (should be iterable)
-        basedir: basedir
-        datasetname: dataset name
-        label: label
-        nmax: nmax in case of subdir
-        minmax_node_feature: minmax_node_feature
-        minmax_graph_feature: minmax_graph_feature
-        comm: MPI communicator
-        """
-        if not os.path.exists(basedir):
-            os.makedirs(basedir)
-        basename = "%s-%s.pkl" % (datasetname, label)
-        fname = os.path.join(basedir, basename)
-        with open(fname, "wb") as f:
-            pickle.dump(minmax_node_feature, f)
-            pickle.dump(minmax_graph_feature, f)
-            pickle.dump(dataset, f)
